+++ lib/mtbuild/test_application_configuration.rb
@@ -17,11 +17,13 @@ module MTBuild
         all_object_folders |= object_folders
       end
 
-      application_binaries, application_files, application_folders = @default_toolchain.create_application_tasks(all_object_files, @parent_project.project_name)
+      project_filename = @parent_project.project_name.to_s.gsub(':','-')
+      application_binaries, application_files, application_folders = @default_toolchain.create_application_tasks(all_object_files, project_filename)
       dependencies = @dependencies+all_object_folders+application_folders+application_files+application_binaries
 
       desc "Build and run test application '#{@parent_project.project_name}' with configuration '#{@configuration_name}'"
       new_task = test_application_task @configuration_name => dependencies do |t|
+        @post_build.call if @post_build.respond_to? :call
         puts "built test application #{t.name}."
         sh "\"#{application_binaries.first}\""
         puts "ran test application #{t.name}."